# api: add support for a new compact float64 representation