@timbray 2024-07-19
Thanks for this!  I'm in the middle of implementing `equals-ignore-case` matching, which turns out to be a messy tangle of corner cases and Unicode weirdness. So I'm going to hold off on seriously studying this until I get that landed.  I wonder if there is any cost of moving from Go19 to 20? Hey @embano1, got an opinion?  I see that some of the unit tests failed, I suppose this is the 19/20 problem.