@timbray 2024-07-19
One note to myself or whoever… we are going to need a decent benchmark to exhibit the runtime performance impact of moving from Q numbers to numbits.  Right now our test that makes heaviest use of numeric matching is `Benchmark_JsonFlattner_Evaluate_ContextFields`.