@Merovius 2024-07-20
ISTM the only reason to use Go 1.20 here is that you use `cmp.Compare` - and only in a test, at that? I'll note that you could also copy [the ten lines or so](https://cs.opensource.google/go/go/+/refs/tags/go1.22.5:src/cmp/cmp.go;l=40) (potentially behind a build tag, with post Go 1.20 versions delegating to the standard library version).