@timbray 2024-07-21
OK.  The bad news is that, while the API is nice, Quamina needs very little of it.  My thinking now is that when flattenJSON encounters a number, we will still use `strconv.ParseDouble()`, which is quite optimized.  Then I would use `NumbitsFromFloat64().Bytes()` to get 8 (non-UTF8) bytes and try to run the automaton on that.  Does that make sense?