@timbray 2024-08-09
OK, I have cloned Arne's branch and will replace qNumber with numbits and do some benchmarking.

1. I like numbits and it would be good if numeric comparisons had real 64-bit accuracy
2. Including numbits is going to complexify some central code paths, especially `smallTable`, which rely on an assumption that all the bytes are UTF-8 bytes and thus can't have values greater than 0xF4, allowing 0xF5 and 0xF6 to be used as sentinels and end-of-data markers.  So, research goal 1: Determine how much complexity is added.
3. Assuming that the numbits basically work and don't lead to horrible complexifying, I'm going to create a benchmark with patterns that include multiple numeric matches, so we understand what we win/lose. My prediction is that numbits should be noticeably faster and burn less memory simply because they are 8 bytes long and qNumbers are 14, but I'm too old to believe in my own predictions. 
4. The result of all this should show us the costs and benefits of numbits and allow us to make an evidence-based decision on whether to adopt.