@timbray 2024-08-18
FYI: I have finished up replacing Quamina's 14-hex-byte Qnumber construct with 10-byte numbits+base12. I still have a bunch of work to do on the README and docs before I can PR, but here are some numbers. I created a new benchmark, `BenchmarkNumberMatching` that simple number mathcing, of which half succeed. The proportion that succeeds is crucial, because to match, the code has to go through the whole 10 bytes to verify all match. 

Using the Go benchmarking, here are before and after results. 
Before:
```
BenchmarkNumberMatching-12    	 1586095	       741.6 ns/op	    2810 B/op	       6 allocs/op
BenchmarkNumberMatching-12    	 1541496	       752.8 ns/op	    2810 B/op	       6 allocs/op
BenchmarkNumberMatching-12    	 1554729	       746.1 ns/op	    2810 B/op	       6 allocs/op
```
After:
```
BenchmarkNumberMatching-12    	 2164224	       537.1 ns/op	    1596 B/op	       6 allocs/op
BenchmarkNumberMatching-12    	 2162497	       537.5 ns/op	    1596 B/op	       6 allocs/op
BenchmarkNumberMatching-12    	 2144856	       548.9 ns/op	    1596 B/op	       6 allocs/op
```
So, that's pretty good!